requirements.txt
fastapi
uvicorn
psycopg2-binary
redis
requests
pydantic
python-multipart
python-dotenv